# openapi.yaml
openapi: 3.0.0

info:
  title: ACME API
  description: |
    ACME Corporation API for Widgets and Doodads.
  version: 1.0.0

servers:
  - url: 'http://localhost:8080'

paths:
  /api/v1/doodads:
    post:
      summary: Create a new doodad
      description: Creates a new doodad with the provided data
      operationId: createDoodad
      requestBody:
        $ref: '#/components/requestBodies/DoodadCreateRequest'
      responses:
        '201':
          $ref: '#/components/responses/DoodadResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntityResponse'
        '500':
          $ref: '#/components/responses/ServerErrorResponse'
    get:
      summary: Get all doodads
      description: Returns a paginated list of all doodads with optional filtering
      operationId: getAllDoodads
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Filter doodads by name (case-insensitive, partial match)
        - name: description
          in: query
          required: false
          schema:
            type: string
          description: Filter doodads by description (case-insensitive, partial match)
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Page number (0-based)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: Page size (default 10, max 100)
      responses:
        '200':
          $ref: '#/components/responses/DoodadPageResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntityResponse'
        '500':
          $ref: '#/components/responses/ServerErrorResponse'

  /api/v1/doodads/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
        description: The doodad ID
    get:
      summary: Get a doodad by ID
      description: Returns a doodad by its ID
      operationId: getDoodadById
      responses:
        '200':
          $ref: '#/components/responses/DoodadResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntityResponse'
        '500':
          $ref: '#/components/responses/ServerErrorResponse'
    put:
      summary: Update a doodad
      description: Updates an existing doodad with the provided data
      operationId: updateDoodad
      requestBody:
        $ref: '#/components/requestBodies/DoodadUpdateRequest'
      responses:
        '200':
          $ref: '#/components/responses/DoodadResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntityResponse'
        '500':
          $ref: '#/components/responses/ServerErrorResponse'

  /api/v1/widgets:
    post:
      summary: Create a new widget
      description: Creates a new widget with the provided data
      operationId: createWidget
      requestBody:
        $ref: '#/components/requestBodies/WidgetCreateRequest'
      responses:
        '201':
          $ref: '#/components/responses/WidgetResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntityResponse'
        '500':
          $ref: '#/components/responses/ServerErrorResponse'
    get:
      summary: Get all widgets
      description: Returns a paginated list of all widgets with optional filtering
      operationId: getAllWidgets
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Filter widgets by name (case-insensitive, partial match)
        - name: description
          in: query
          required: false
          schema:
            type: string
          description: Filter widgets by description (case-insensitive, partial match)
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Page number (0-based)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: Page size (default 10, max 100)
      responses:
        '200':
          $ref: '#/components/responses/WidgetPageResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntityResponse'
        '500':
          $ref: '#/components/responses/ServerErrorResponse'

  /api/v1/widgets/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
        description: The widget ID
    get:
      summary: Get a widget by ID
      description: Returns a widget by its ID
      operationId: getWidgetById
      responses:
        '200':
          $ref: '#/components/responses/WidgetResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntityResponse'
        '500':
          $ref: '#/components/responses/ServerErrorResponse'
    put:
      summary: Update a widget
      description: Updates an existing widget with the provided data
      operationId: updateWidget
      requestBody:
        $ref: '#/components/requestBodies/WidgetUpdateRequest'
      responses:
        '200':
          $ref: '#/components/responses/WidgetResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntityResponse'
        '500':
          $ref: '#/components/responses/ServerErrorResponse'


components:
  schemas:
    PageInfo:
      type: object
      description: Information about a page of results
      properties:
        number:
          type: integer
          description: The current page number (0-based)
        size:
          type: integer
          description: The size of the page
        totalPages:
          type: integer
          description: The total number of pages
        totalElements:
          type: integer
          format: int64
          description: The total number of elements across all pages

    DoodadPageResponse:
      type: object
      description: A paginated response containing doodads
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/DoodadDto'
          description: The list of doodads in the current page
        page:
          $ref: '#/components/schemas/PageInfo'
          description: Pagination information

    WidgetPageResponse:
      type: object
      description: A paginated response containing widgets
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/WidgetDto'
          description: The list of widgets in the current page
        page:
          $ref: '#/components/schemas/PageInfo'
          description: Pagination information

    DoodadDto:
      type: object
      description: Data Transfer Object for Doodad entities
      properties:
        id:
          type: integer
          format: int64
          description: The doodad identifier
        name:
          type: string
          description: The doodad name
        description:
          type: string
          description: The doodad description
        image:
          type: string
          description: The doodad image reference
        price:
          type: number
          format: float
          description: The doodad price

    DoodadRequest:
      type: object
      description: Data Transfer Object for creating or updating Doodad entities
      required:
        - name
        - description
        - image
        - price
      properties:
        name:
          type: string
          description: The doodad name
        description:
          type: string
          description: The doodad description
        image:
          type: string
          description: The doodad image reference
        price:
          type: number
          format: float
          description: The doodad price

    WidgetDto:
      type: object
      description: Data Transfer Object for Widget entities
      properties:
        id:
          type: integer
          format: int64
          description: The widget identifier
        name:
          type: string
          description: The widget name
        description:
          type: string
          description: The widget description
        image:
          type: string
          description: The widget image reference

    WidgetRequest:
      type: object
      description: Data Transfer Object for creating or updating Widget entities
      required:
        - name
        - description
        - image
      properties:
        name:
          type: string
          description: The widget name
        description:
          type: string
          description: The widget description
        image:
          type: string
          description: The widget image reference

    ProblemDetail:
      type: object
      description: Standard error response following RFC 7807
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          description: A URI reference that identifies the problem type
          example: "https://example.com/problems/constraint-violation"
        title:
          type: string
          description: A short, human-readable summary of the problem type
          example: "Constraint Violation"
        status:
          type: integer
          description: The HTTP status code
          example: 400
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem
          example: "The name field cannot be empty"
        instance:
          type: string
          description: A URI reference that identifies the specific occurrence of the problem
          example: "/api/v1/widgets/123"
        properties:
          type: object
          description: Additional properties specific to the error
          additionalProperties: true

  requestBodies:
    DoodadCreateRequest:
      description: Request body for creating a doodad
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DoodadRequest'

    DoodadUpdateRequest:
      description: Request body for updating a doodad
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DoodadRequest'

    WidgetCreateRequest:
      description: Request body for creating a widget
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WidgetRequest'

    WidgetUpdateRequest:
      description: Request body for updating a widget
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WidgetRequest'

  responses:
    DoodadResponse:
      description: A single doodad
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DoodadDto'

    DoodadPageResponse:
      description: A paginated list of doodads
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DoodadPageResponse'

    WidgetResponse:
      description: A single widget
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WidgetDto'

    WidgetPageResponse:
      description: A paginated list of widgets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WidgetPageResponse'

    BadRequestResponse:
      description: Bad request due to invalid input
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            validation-error:
              value:
                type: "https://example.com/problems/validation-error"
                title: "Validation Error"
                status: 400
                detail: "The request contains invalid parameters"
                instance: "/api/v1/resources"
            widget-validation-error:
              value:
                type: "https://example.com/problems/validation-error"
                title: "Validation Error"
                status: 400
                detail: "The request contains invalid parameters"
                instance: "/api/v1/widgets"
            doodad-validation-error:
              value:
                type: "https://example.com/problems/validation-error"
                title: "Validation Error"
                status: 400
                detail: "The request contains invalid parameters"
                instance: "/api/v1/doodads"

    NotFoundResponse:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            resource-not-found:
              value:
                type: "https://example.com/problems/not-found"
                title: "Resource Not Found"
                status: 404
                detail: "The requested resource could not be found"
                instance: "/api/v1/resources/999"
            widget-not-found:
              value:
                type: "https://example.com/problems/not-found"
                title: "Resource Not Found"
                status: 404
                detail: "The requested widget could not be found"
                instance: "/api/v1/widgets/999"
            doodad-not-found:
              value:
                type: "https://example.com/problems/not-found"
                title: "Resource Not Found"
                status: 404
                detail: "The requested doodad could not be found"
                instance: "/api/v1/doodads/999"

    UnprocessableEntityResponse:
      description: Unprocessable entity due to validation errors
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            validation-error:
              value:
                type: "https://example.com/problems/validation-error"
                title: "Validation Error"
                status: 422
                detail: "The request contains validation errors"
                instance: "/api/v1/widgets"
                properties:
                  errors: 
                    name: "Name must be between 3 and 50 characters"
            doodad-validation-error:
              value:
                type: "https://example.com/problems/validation-error"
                title: "Validation Error"
                status: 422
                detail: "The request contains validation errors"
                instance: "/api/v1/doodads"
                properties:
                  errors: 
                    name: "Name must be between 3 and 50 characters"
                    price: "Price must be a positive number"

    ServerErrorResponse:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            server-error:
              value:
                type: "https://example.com/problems/server-error"
                title: "Internal Server Error"
                status: 500
                detail: "An unexpected error occurred"
                instance: "/api/v1/resources"
            widget-server-error:
              value:
                type: "https://example.com/problems/server-error"
                title: "Internal Server Error"
                status: 500
                detail: "An unexpected error occurred"
                instance: "/api/v1/widgets"
            doodad-server-error:
              value:
                type: "https://example.com/problems/server-error"
                title: "Internal Server Error"
                status: 500
                detail: "An unexpected error occurred"
                instance: "/api/v1/doodads"